# feat!: support multiple parallel `child_process`