--- a/packages/vitest/src/integrations/mock/timers.ts
+++ b/packages/vitest/src/integrations/mock/timers.ts
@@ -134,6 +134,10 @@ export class FakeTimers {
         // Do not mock nextTick by default. It can still be mocked through userConfig.
         .filter(timer => timer !== 'nextTick') as (keyof FakeTimerWithContext['timers'])[]
 
+      // @ts-expect-error -- untyped internal
+      if (this._userConfig?.toFake?.includes('nextTick') && globalThis.__vitest_worker__.isChildProcess)
+        throw new Error('process.nextTick cannot be mocked inside child_process')
+
       this._clock = this._fakeTimers.install({
         now: Date.now(),
         toFake,