--- a/packages/vitest/src/types/child.ts
+++ b/packages/vitest/src/types/child.ts
@@ -1,5 +1,5 @@
 import type { ContextRPC } from './rpc'
 
 export interface ChildContext extends ContextRPC {
-  command: 'start'
+  workerId: number
 }