--- a/packages/vitest/src/types/worker.ts
+++ b/packages/vitest/src/types/worker.ts
@@ -35,4 +35,5 @@ export interface WorkerGlobalState {
     environment: number
     prepare: number
   }
+  isChildProcess?: boolean
 }