--- a/packages/vitest/src/types/worker.ts
+++ b/packages/vitest/src/types/worker.ts
@@ -4,7 +4,7 @@ import type { ModuleCacheMap, ViteNodeResolveId } from 'vite-node'
 import type { BirpcReturn } from 'birpc'
 import type { MockMap } from './mocker'
 import type { ResolvedConfig } from './config'
-import type { ContextRPC, RuntimeRPC } from './rpc'
+import type { ContextRPC, RunnerRPC, RuntimeRPC } from './rpc'
 import type { Environment } from './general'
 
 export interface WorkerContext extends ContextRPC {
@@ -18,7 +18,7 @@ export interface AfterSuiteRunMeta {
   coverage?: unknown
 }
 
-export type WorkerRPC = BirpcReturn<RuntimeRPC>
+export type WorkerRPC = BirpcReturn<RuntimeRPC, RunnerRPC>
 
 export interface WorkerGlobalState {
   ctx: ContextRPC