--- a/test/config/fixtures/test/fake-timers.test.ts
+++ b/test/config/fixtures/test/fake-timers.test.ts
@@ -1,0 +1,7 @@
+import { test, vi } from 'vitest'
+
+test('uses fake timers', () => {
+  vi.useFakeTimers()
+
+  vi.useRealTimers()
+})