… docs/config/index.md
… docs/guide/debugging.md
… package.json
… packages/ui/package.json
… packages/vitest/package.json
… packages/ws-client/package.json
… pnpm-lock.yaml