### Description

- Many projects are unable to use `node:worker_threads` and are forced to use a single `node:child_process` via `--no-threads`. That's stable, but slow. This PR allows using multiple `node:child_process` for running tests similarly as `--threads` does.
  - https://github.com/vitest-dev/vitest/issues/3143
  - https://github.com/vitest-dev/vitest/issues/3106
  - https://github.com/vitest-dev/vitest/issues/3816
  - https://github.com/tinylibs/tinypool/issues/48
- Ref. https://github.com/tinylibs/tinypool/issues/64

#### BREAKING CHANGES

`useFakeTimers()` no longer automatically mocks `process.nextTick`. Users can still mock it by explicitly specifying `useFakeTimers({ toFake: ['nextTick'] })`.

Mocking `process.nextTick` by using `useFakeTimers({ toFake: ['nextTick'] })` is no longer supported when running in `--no-threads`. Node's `child_process` uses `process.nextTick` internally and it is really difficult to support this use case. Users can still mock `process.nextTick` when running with `--threads`.

`--no-threads` no longer runs tests in the same test context. Tests are now isolated in separate processes similar to how `--threads` does with `worker_threads`.

#### Testing

 - [x] `pnpm test:ci:no-threads`
 - Manual testing with third party projects:
   - [x] [`zustand`](https://github.com/pmndrs/zustand): Tests require isolation so comparing speed change doesn't make sense here. All tests pass with new isolated `--no-threads`.
   - [x] [`vuejs/core`](https://github.com/vuejs/core). Uses `--no-threads` on CI. Speed improvement 48.57s -> 26.96s.
   - [x] [`netlify/cli`](https://github.com/netlify/cli). Tests pass with `--no-threads`.
   - [x] [`lerna-lite`](https://github.com/lerna-lite/lerna-lite). Tests require `child_process` and have been using `--no-threads`. Speed improvement 67s -> 24s.

___

### Please don't delete this checklist! Before submitting the PR, please make sure you do the following:
- [ ] It's really useful if your PR references an issue where it is discussed ahead of time. If the feature is substantial or introduces breaking changes without a discussion, PR might be closed.
- [x] Ideally, include a test that fails without this PR but passes with it.
- [x] Please, don't make changes to `pnpm-lock.yaml` unless you introduce a new test example.

### Tests
- [x] Run the tests with `pnpm test:ci`.

### Documentation
- [x] If you introduce new functionality, document it. You can run documentation with `pnpm run docs` command.

### Changesets
- [x] Changes in changelog are generated from PR name. Please, make sure that it explains your changes in an understandable manner. Please, prefix changeset messages with `feat:`, `fix:`, `perf:`, `docs:`, or `chore:`.