@AriPerkkio 2023-08-29
Seeing some process hangs when running Vitest's test cases with `--no-threads`. Debugging these is really time consuming and I've spent quite some time looking into this. I'm starting to think we should implement graceful `worker_thread` and `child_process` termination in Tinypool. That would allow us to implement Jest's `--detect-open-handles` feature too.

These errors come up locally and on CI:

```
⎯⎯⎯⎯⎯⎯ Unhandled Error ⎯⎯⎯⎯⎯⎯⎯
Error: [birpc] timeout on calling "onTaskUpdate"
 ❯ Timeout._onTimeout ../../packages/vitest/dist/vendor-index.9acf798a.js:38:22
 ❯ listOnTimeout node:internal/timers:569:17
 ❯ processTimers node:internal/timers:512:7
```