@sheremet-va 2023-08-30
> Debugging these is really time consuming and I've spent quite some time looking into this. 

Welcome to my world! For me it's almost always something related to timer mocks (nextTick included).

> I'm starting to think we should implement graceful worker_thread and child_process termination in Tinypool.

Will it help? Or just to debug?