@AriPerkkio 2023-08-30
> Will it help? Or just to debug?

Mostly for debugging. But still, I'm not sure if it's good practice to terminate workers and processess with `.kill()` and `.terminate()` when there can still be all kind of stuff running, e.g. chokidar with fsevents. I'm starting to think that it's causing some of these process hangs but don't really have proof for it.

The way how `jest-worker` does this is that main thread instructs thread/process (worker) to [remove all its message handlers](https://github.com/jestjs/jest/blob/25a8785584c9d54a05887001ee7f498d489a5441/packages/jest-worker/src/workers/processChild.ts#L155C52-L156) so that there is nothing keeping the worker running, and then it waits for the worker to gracefully exit. If it doesn't exit in specific timeout, main threads sends it SIGTERM. If it still doesn't exit, [it sends SIGKILL](https://github.com/jestjs/jest/blob/25a8785584c9d54a05887001ee7f498d489a5441/packages/jest-worker/src/workers/ChildProcessWorker.ts#L463C8-L477). What Tinypool currently does is that it immediately sends SIGTERM and then hangs if worker doesn't exit - or throws [timeout error](https://github.com/tinylibs/tinypool/pull/50).

But if Vitest used graceful termination for workers, it could impact in performance. We would not be killing workers as fast as possible - instead we would be waiting for them to shut down themselves. I'm not sure if this performance issue is real or just speculation. Jest or Vitest's `vm-threads` don't really suffer from this as isolation doesn't require recycling workers.

Maybe I'll implement `--detect-open-handles` into Tinypool so that it's optional flag that can be turned on when needed. It would definitely help debugging this PR's issues.