@ghiscoding 2023-09-13
> * [`lerna-lite`](https://github.com/lerna-lite/lerna-lite). Tests require `child_process` and have been using `--no-threads`. Speed improvement 67s -> 24s. Some snapshots are wrapped in double quotes now. 🤔

@AriPerkkio I'm really excited about this PR, especially that I'm on Windows so this will be a big boost for me... and I just saw that you started testing Lerna-Lite, awesome, so about the comment you wrote, I might be able to shed some light, I had a few issues when I first moved to Vitest (like this issue #3129 that I logged in Vitest). Basically Lerna (Lerna-Lite is a fork of Lerna) had many custom serializers and running them with `--no-threads` had the consequence of being run in the same global scope causing some problems as described by @sheremet-va in the referenced issue 

> All tests with the same environment in `--no-threads` runs in a single child_process and they all share the same global scope. Is it possible that one test just affects the other?

I fixed this problem in this Lerna-Lite [PR](https://github.com/lerna-lite/lerna-lite/pull/650) by duplicating some of these serializer in certain spec files that had quote escaping issues

So in short, it might be related to the double quotes problem you are encountering when testing Lerna-Lite. I don't mind reverting this Lerna-Lite [PR](https://github.com/lerna-lite/lerna-lite/pull/650) or changing any code that could help with the extra quotes you're seeing, whatever makes it works, I'm all good with it 😉