@AriPerkkio 2023-09-13
Thanks for the insight @ghiscoding - you just save me an hour or two of debugging. In the current testing setup of `lerna-lite` the snapshot serializers are indeed leaking from one test file to another. I fixed the snapshot issues by adding the serializer to following files: `version-build-metadata.spec.ts`, `version-conventional-commits.spec.ts`, `version-ignore-changes.spec.ts`. Now all tests are passing and running quickly.