@AriPerkkio 2023-09-13
What kind of API should `child_process` pool expose to users, @sheremet-va?

- `--single-process`, `minProcess`, `maxProcess` (equivalents of `--single-thread`, `minThreads`, `maxThreads`
- Or simply call threads and processes as "workers" as Jest does, and have `--single-worker`, `minWorkers`, `maxWorkers`? Pool selection would be done using `--threads`, `--no-threads`? 

Currenly this PR does not implement a `child_process` equivalent of `--single-threads` - do we need one?