@sheremet-va 2023-09-13
The problem here is naming 😄

Maybe implement current names for bow, but we already discussed the new naming convention for pools - maybe we can implement it in the same breaking change release, but in a separate PR.