@ghiscoding 2023-09-13
> Thanks for the insight @ghiscoding - you just saved me an hour or two of debugging. In the current testing setup of `lerna-lite` the snapshot serializers are indeed leaking from one test file to another. I fixed the snapshot issues by adding the serializer to following files: `version-build-metadata.spec.ts`, `version-conventional-commits.spec.ts`, `version-ignore-changes.spec.ts`. Now all tests are passing and running quickly.

I would be glad to receive a PR with all your changes for this upcoming feature, even if it's in draft until this is released, that would save me some time when it becomes available. Thanks a ton for all your hard work :)