@AriPerkkio 2023-09-14
> Maybe implement current names for now, but we already discussed the new naming convention for pools - maybe we can implement it in the same breaking change release, but in a separate PR.

Yes, now I remember it - again. 😄 Configuring via pools make this much easier. 

I'll leave this PR to use `--no-threads` and implement `--single-thread` equivalent for `child_process`. I don't think there's a way to use that without introducing new configuration option but let's leave that to another PR.