@AriPerkkio 2023-09-14
> [...] and implement `--single-thread` equivalent for `child_process`. I don't think there's a way to use that without introducing new configuration option but let's leave that to another PR.

Looks like I already implemented this. It's available with `vitest --no-threads --single-thread`. The `--no-thread` makes Vitest pick `pools/child.ts` and `--single-thread` makes it run in a single process. 

> #### TODO
>
> - [ ] Reporters work weird in test/public-api and test/stacktraces when run with --no-threads. Enabling CI=true fixes this.

Looks like this bug is present on `main` too. When running Vitest through JS API (`startVitest()`) with `--no-threads`, the reporters are not resolved correctly. This can be tested with:

```sh
cd vitest/test/public-api

# Pass
pnpm run test --threads --run
CI=true pnpm run test --no-threads --run

# Fail
pnpm run test --no-threads --run
```