@sheremet-va 2023-09-15
> so is this complete? on which milestone should we expect this to land, in v0.35 or v1.0? As you can see I'm very exited about it 😄

The next major is 1.0.0-beta.0