@ghiscoding 2023-09-15
> The next major is 1.0.0-beta.0

@sheremet-va ok but just so you know, you still have a v0.35 milestone that exist with 1 PR assign to it 😋 
https://github.com/vitest-dev/vitest/milestone/11

awesome work guys